requirements.txt
Django==3.0.2
django-gmailapi-backend==0.1.0
django-sesame==2.1
google-api-python-client==1.7.11
google-auth==1.11.2
google-auth-httplib2==0.0.4
Pillow==7.0.0
